feat: Align EmbeddingOptions builder to ChatOptions
* Add Builder interface under EmbeddingOptions, and provide default implementation with DefaultEmbeddingOptions and DefaultEmbeddingOptionsBuilder. This is to match the design adopted alredy for ChatOptions in 1.0.0
